replaced completely old scale database with Room database included migrating method

+++ android_app/app/src/main/java/com/health/openscale/gui/activities/DataEntryActivity.java
@@ -464,7 +464,7 @@ public class DataEntryActivity extends Activity {
         }
 
         void deleteMeasurement() {
-            long delId = id;
+            int delId = id;
 
             boolean hasNext = moveLeft();
